# worker/openapi.yaml
openapi: 3.0.3
info:
  title: E-Ink Board Worker API
  version: "1.0"
  description: |
    API for the Cloudflare Worker backing the EE02 e-ink display boards.

    Two very different kinds of clients call this API:

    - **Admin clients** (you, from a browser, curl, Postman, etc.) — everything
      under `/admin/*` plus `GET /current`. Authenticate with
      `Authorization: Bearer <api_key>`. Get a key by completing passkey
      registration (`/auth/register/*`) or by copying it from the admin UI at
      `/admin` ("Rotate API Key").
    - **Devices** (the ESP32 firmware) — `/device_config`, `/hash`,
      `/image_packed`, `/firmware_bin`. These use a different, HMAC-based
      per-request signature (`X-Device-MAC` / `X-Device-Nonce` /
      `X-Device-Signature`), not the Bearer key. Documented here for
      completeness but you won't normally call these by hand.

    Import this file into Postman/Insomnia/Swagger UI, set the `bearerAuth`
    security value to your API key, and every `/admin/*` request (including
    `GET /current`) is one click away.
servers:
  - url: https://{host}
    variables:
      host:
        default: localhost:8787
        description: Your deployed worker's hostname (or a local `wrangler dev` host:port).
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: >-
        Your account's API key, minted by /auth/register/verify,
        /auth/login/verify, or POST /admin/keys/rotate.
  schemas:
...
    Image:
      type: object
      properties:
        id:
          type: string
        filename:
          type: string
          maxLength: 255
          description: No control characters (it's echoed in the X-Image-Name response header).
        dither_algorithm:
          type: string
          enum: [floyd_steinberg, atkinson, ordered, none]
        raw_bytes:
          type: integer
        created_at:
          type: integer
        key_version:
          type: integer
          description: >-
            Which of the bucket's key versions this image's raw blob and
            every board variant's blobs are actually encrypted under right
            now — see migrations/0016_bucket_key_rotation.sql. Lags the
            bucket's own key_version between a rotation's start and this
            image's reencrypt-image call.
        variants:
          type: object
          description: >-
            A bucket is never board-scoped — see
            migrations/0019_image_board_variants.sql. Keyed by board id
            (e.g. `ee02-13in3`, `ee04-7in3`); every upload generates every
            board's variant, so this is normally populated for all of them.
          additionalProperties:
            type: object
            properties:
              packed_hash: { type: string }
              packed_bytes: { type: integer }
              packed_encoding: { type: string, enum: [identity, deflate-raw] }
        thumbnail_ciphertext_b64:
          type: string
          nullable: true
          description: >-
            Base64 of the AES-256-GCM-encrypted thumbnail (12-byte nonce
            prepended) for one representative board's variant (a human
            preview, not board-specific). The dashboard decrypts this
            client-side with the bucket key — it is not a renderable data
            URL on its own.
...
  parameters:
...
    ImageIdPath:
      name: id
      in: path
      required: true
      schema: { type: string }
...
paths:
...
  /admin/images:
    get:
      summary: List a bucket's images
      description: >-
        Read-shaped access: allowed for the owner, an accepted collaborator,
        OR anyone at all when the bucket is public
        (migrations/0018_public_buckets.sql) — see
        lib/bucket-access.ts's assertBucketReadAccess.
      parameters:
        - name: device_key
          in: query
          required: true
          description: The bucket id.
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  images:
                    type: array
                    items: { $ref: '#/components/schemas/Image' }
        '403': { description: No read access to this bucket }
  /admin/images/upload:
    post:
      summary: Upload an already-encrypted image into a bucket
      description: >-
        Decode, EXIF-correct, crop, enhance, dither, pack, and AES-256-GCM-encrypt
        all run client-side now — see root CLAUDE.md's encrypted-buckets plan. This
        endpoint never sees plaintext: it stores one opaque raw-original ciphertext
        blob plus, for every board (migrations/0019_image_board_variants.sql — a
        bucket is never board-scoped), that board's packed+thumbnail ciphertext and
        a client-computed content hash, gated by the same bucket-access check as
        before. `dither_algorithm` and each board's `packed_hash` are trusted
        client-reported metadata — the Worker has no way to verify them without the
        bucket key.
      parameters:
        - name: device_key
          in: query
          required: true
          description: The bucket id to upload into.
          schema: { type: string }
        - name: filename
          in: query
          required: true
          schema: { type: string }
        - name: allow_duplicate
          in: query
          required: false
          description: >-
            Set to 1 to bypass the 409 duplicate rejection below (the client
            sets this when the user confirms "upload anyway").
          schema: { type: string, enum: ['1'] }
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - raw
                - packed__ee02-13in3
                - thumb__ee02-13in3
                - packed_hash__ee02-13in3
                - packed__ee04-7in3
                - thumb__ee04-7in3
                - packed_hash__ee04-7in3
              properties:
                raw: { type: string, format: binary, description: AES-256-GCM ciphertext (12-byte nonce prepended) of the original as-uploaded bytes, shared by every board's variant. }
                dither_algorithm:
                  type: string
                  enum: [floyd_steinberg, atkinson, ordered]
                  default: floyd_steinberg
                packed__ee02-13in3: { type: string, format: binary, description: AES-256-GCM ciphertext of EE02's packed 4bpp buffer (1600x1200) }
                thumb__ee02-13in3: { type: string, format: binary, description: AES-256-GCM ciphertext of EE02's dashboard preview JPEG }
                packed_hash__ee02-13in3: { type: string, description: 16-char hex content hash over EE02's packed ciphertext. }
                packed_encoding__ee02-13in3: { type: string, enum: [identity, deflate-raw], default: identity }
                packed__ee04-7in3: { type: string, format: binary, description: AES-256-GCM ciphertext of EE04's packed 4bpp buffer (800x480) }
                thumb__ee04-7in3: { type: string, format: binary, description: AES-256-GCM ciphertext of EE04's dashboard preview JPEG }
                packed_hash__ee04-7in3: { type: string, description: 16-char hex content hash over EE04's packed ciphertext. }
                packed_encoding__ee04-7in3: { type: string, enum: [identity, deflate-raw], default: identity }
                content_hash:
                  type: string
                  description: >-
                    Optional 16-char hex duplicate-detection hash (migrations/
                    0020_image_content_hash.sql): HMAC-SHA256 keyed with a
                    bucket-key-derived hash key (client/crypto.ts's
                    computeContentHash) over the DEFAULT board's plaintext
                    packed buffer, computed before compression/encryption so
                    identical renditions collide regardless of source-file
                    format or filename. Keyed so the Worker can only compare
                    within one bucket — no cross-bucket correlation. Trusted
                    client metadata, never verified; absent = no dedupe for
                    this upload.
      responses:
        '201': { description: Uploaded }
        '400': { description: Missing or malformed field }
        '409':
          description: >-
            Duplicate rendition — the bucket already contains an image with
            this content_hash under a different filename. Body includes
            `duplicate_of` (the existing filename). Retry with
            `?allow_duplicate=1` to force the upload through.
        '403':
          description: >-
            No WRITE access to this bucket — unlike GET /admin/images, being
            public is never enough on its own (see
            migrations/0018_public_buckets.sql: public buckets are
            read-only for everyone but the owner/collaborators).
  /admin/images/{id}:
    delete:
      summary: Delete an image
      parameters:
        - $ref: '#/components/parameters/ImageIdPath'
      responses:
        '200': { description: Deleted }
        '403': { description: No WRITE access to this bucket (see /admin/images/upload) }
        '404': { description: Not found }
  /admin/images/{id}/raw:
    get:
      summary: Fetch an image's original as-uploaded ciphertext
      description: >-
        Read-shaped access — same rule as GET /admin/images: allowed for the
        owner, an accepted collaborator, or anyone at all when the bucket is
        public (migrations/0018_public_buckets.sql).
      parameters:
        - $ref: '#/components/parameters/ImageIdPath'
      responses:
        '200':
          description: OK
          content:
            image/*: { schema: { type: string, format: binary } }
        '403': { description: No read access to this bucket }
        '404': { description: Not found }
...